refactor notification request mapper and re-use get selected court details
